Refactor IngredientService to be more Object Oriented and test
It was desired to make the ingredient and recipe loaders more oo.
This was done however now tests are broken and we cannot commit!

Also noted was the duplication of test data while writing tests which just
feels wrong. Time to bring in the mocks and supercharge our tests.

I also noted that sometimes the "expected" and "actual" parameters were switched
so I've fixed any observed instances of this.